initial attempt at emailing
