+++ cs251tk/formatters/__init__.py
@@ -1,5 +1,4 @@
 from .markdown import format_assignment_markdown as markdown
 from .gist import format_assignment_gist as gist
 from .yaml import format_assignment_yaml as yaml
-# from .email import format_collected_data as email
 from .base import format_collected_data
