+++ cs251tk/referee/__init__.py
@@ -1,2 +1,6 @@
 from .args import process_args
 from .process_student import process_student
+from .parse_commits import parse_commits_for_assignments
+from .emailify import emailify
+from .send_email import send_email
+from .send_recordings import send_recordings
