+++ cs251tk/referee/send_email.py
@@ -1,0 +1,8 @@
+import smtplib
+
+
+def send_email(msg):
+    # Send the message via our own SMTP server.
+    s = smtplib.SMTP('localhost')
+    s.send_message(msg)
+    s.quit()
